refactor(cat): split tool-use into types/detection/executor (545→182)
tool-use.ts was 545 LOC — types, intent detection, tool definitions, the
single-tool executor, and the agentic orchestration loop all in one file.
Extracted verbatim (behavior-preserving):
- tool-use-types.ts — message/event/proposal types (+ RawToolCall)
- tool-use-detection.ts — messageMightNeedTools, hasCreateIntent,
  PREFILLABLE_ENTITY_TYPES, PLATFORM_TOOL_DEFINITION
- tool-executor.ts — executeToolCall (search_platform + prefill_entity_form)

tool-use.ts (182 LOC) keeps the maybeEnrichWithSearchResults orchestration loop
and re-exports the full public surface, so chat-orchestrator's imports are
unchanged.

tsc 0 errors.
